% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{api_to_specific_gravity}
\alias{api_to_specific_gravity}
\title{Convert API to specific gravity}
\usage{
api_to_specific_gravity(api)
}
\arguments{
\item{api}{gravity of oil or condensate in API degrees}
}
\description{
Convert API to specific gravity
}
