% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{calc_temperature_gradient}
\alias{calc_temperature_gradient}
\title{Calculate the temperature gradient based on bottomhole and wellhead temperature}
\usage{
calc_temperature_gradient(bht, wht, depth.bh)
}
\arguments{
\item{bht}{bottomhole temperature, deg F}

\item{wht}{wellhead temperature, deg F}
}
\description{
Calculate the temperature gradient based on bottomhole and wellhead temperature
}
