% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_extdata_dir}
\alias{get_extdata_dir}
\title{Get the 'extdata' folder in a package}
\usage{
get_extdata_dir(pkgname)
}
\arguments{
\item{pkgname}{the name of the package}
}
\description{
Get the 'extdata' folder in a package
}
