% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_list_examples}
\alias{get_list_examples}
\title{Get a list of R scripts under examples that end in App}
\usage{
get_list_examples(folder = "examples", aPackage, ...)
}
\arguments{
\item{folder}{folder where the examples live}

\item{aPackage}{package to analyze}

\item{...}{additional parameters}
}
\description{
Get a list of R scripts under examples that end in App
}
