% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{importFromExamples}
\alias{importFromExamples}
\title{Source an R script from inst/examples}
\usage{
importFromExamples(aClassFile, aFolder = "examples", aPackage)
}
\arguments{
\item{aClassFile}{a file containing one or more classes}

\item{aFolder}{a folder where examples are located}

\item{aPackage}{a package where the folder and class resides}
}
\description{
This is useful when we separate an application or implementation from the
class. Similar to what OpenSourcePhysics does: the ODE classes in separate
files and the main application in classApp
}
