% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_from_project}
\alias{load_from_project}
\title{Load an object from an rda file}
\usage{
load_from_project(name = NULL)
}
\arguments{
\item{name}{the name of the dataset}
}
\description{
Load the file to the project inst/extdata folder
}
