% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_csv_from_project}
\alias{read_csv_from_project}
\title{Read a CSV file from a project folder inst/extdata
It saves time from specifying path and everything}
\usage{
read_csv_from_project(name = NULL)
}
\arguments{
\item{name}{the name of the dataset, no extension}
}
\description{
Read a CSV file from a project folder inst/extdata
It saves time from specifying path and everything
}
