% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_csv_to_project}
\alias{save_csv_to_project}
\title{Save a CSV file to a project folder inst/extdata
It saves time from specifying path and everything}
\usage{
save_csv_to_project(object, name = NULL)
}
\arguments{
\item{object}{it is R object to save}

\item{name}{the name of the dataset, no extension}
}
\description{
Save a CSV file to a project folder inst/extdata
It saves time from specifying path and everything
}
