% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_to_project}
\alias{save_to_project}
\title{Save an object to an rda file}
\usage{
save_to_project(..., name = NULL)
}
\arguments{
\item{...}{any object}

\item{name}{the name of the dataset}
}
\description{
Saves the file to the project onst/extdata folder
}
