% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{specific_gravity_to_api}
\alias{specific_gravity_to_api}
\title{Convert Speciific Gravity of oil or condensate to API}
\usage{
specific_gravity_to_api(sg)
}
\arguments{
\item{sg}{specific gravity of oil or condensate}
}
\description{
Convert Speciific Gravity of oil or condensate to API
}
